fix(ui): simplify datepicker structure
