+++ .changeset/modern-oranges-roll.md
@@ -1,0 +1,5 @@
+---
+'@lion/ui': patch
+---
+
+datepicker uses no calender-overlay-frame element anymore
