--- packages/ui/exports/define/lion-calendar-overlay-frame.js
@@ -1,3 +1,0 @@
-import { LionCalendarOverlayFrame } from '../input-datepicker.js';
-
-customElements.define('lion-calendar-overlay-frame', LionCalendarOverlayFrame);
