+++ packages/ui/exports/input-datepicker.js
@@ -1,2 +1,1 @@
 export { LionInputDatepicker } from '../components/input-datepicker/src/LionInputDatepicker.js';
-export { LionCalendarOverlayFrame } from '../components/input-datepicker/src/LionCalendarOverlayFrame.js';
